rename alignment files
