+++ test/align/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectories()
-seqan3_test(alignment_class_test.cpp)
+seqan3_test(alignment_test.cpp)
